Cast to uint16_t.

+++ src/essence.game/qpang/room/session/npc/RoomSessionNpcManager.cpp
@@ -241,7 +241,7 @@ void RoomSessionNpcManager::onCGPvEHitNpc(const CGPvEHitNpcData& data)
 
 	const auto damage = targetNpc->calculateHitDamage(data);
 
-	const auto damageDealtToBodyPart = targetNpc->takeBodyPartDamage(data.bodyPartId, damage);
+	const auto damageDealtToBodyPart = targetNpc->takeBodyPartDamage(data.bodyPartId, static_cast<uint16_t>(damage));
 	const auto damageDealtToNpc = targetNpc->takeDamage(static_cast<uint16_t>(damage));
 
 	const auto hasTargetDied = targetNpc->isDead();
